Make reverse proxy work with blocks
There were problem in etc/nginx/conf.d/blocks/locations.conf
that caused return of 404 (Not Found) when requested file from
sub dir. This was because of using expires header and they
are not supposed to be used with reverse proxy.

Move those expires to etc/nginx/conf.d/blocks/expires.conf
which should be included if used in local Nginx and not
in reverse proxy.

Added also needed tests for HURL and configuration for
reverse proxy.